requirements/base.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-ckeditor==6.7.0
drf-yasg==1.21.7
python-dotenv==1.0.0
Pillow==10.0.1

